deal with internal buffer growing over 2g in size
this fixes a corner case where you may buffer 2g or more of data
in ecore-con buffers. this leads to a stall.

Backport from commit 16d7b981ebc22df86e4dc7719ae63677bff8b1e1
on https://git.enlightenment.org/core/efl.git

+++ src/lib/ecore_con/ecore_con_local_win32.c
@@ -687,7 +687,7 @@ ecore_con_local_connect(Ecore_Con_Server *svr,
 Eina_Bool
 ecore_con_local_win32_server_flush(Ecore_Con_Server *svr)
 {
-   int num;
+   size_t num;
    BOOL res;
    DWORD written;
 
@@ -734,7 +734,7 @@ Eina_Bool
 ecore_con_local_win32_client_flush(Ecore_Con_Client *cl)
 {
    Ecore_Con_Type type;
-   int num;
+   size_t num;
    BOOL res;
    DWORD written;
